CCAR: change basic types name to have only one name for these types. All basic types had 2 names : one internal (Double, for example) and one external (double, for example). This was not a sane situation. For example, in XML files it is possible to use double and in GUI to use Double but adding an alias Double <-> double. I have kept the external names double, int, bool and string and changed all internal names to these ones. The files modified are Any.cxx, TypeCode.cxx, VisitorSaveSchema.cxx, engineTest.cxx for the non GUI part All the other files are for the GUI part except 2 corrections of bugs in GUI in YACSGui_Observer.cxx, YACSGui_Table.cxx and YACSGui_Graph.*
Bug 1 : when a node has an input port with the same name as an output port,
the update during execution is wrong (output port replaces input port). CORRECTED.

Bug 2 : creating an inlineNode with several ports (output for example)
and executing just after creation results in a node with all output
ports with the same name (method setReadOnly of Table). CORRECTED.