Fix a regression made by the HPContainer development in the GUI (commit 54dd4bca5d0eb605fae21321a06bbdee31dbe886).
Before the HPContainer development, when a node was added to the
schema from the catalog, a new instance of the component was created.
It was possible to change the instance, if desired.

After the HPContainer development, the same instance was used every
time the same service was added to the schema and there was no choice
to use different instances for the same service on different nodes.

This fix goes back to the behaviour before the HPContainer development.