Add py2yacs
py2yacs library can build a yacs schema out of a python script.
The schema contains a single python node which executes a function
from the python script.
Input and output ports are deducted from the parameters and the
returns of the chosen function. The type of ports are restricted
to 'double'.